die horribly with null request id
